Redo player progress reporting back to highly granular

+++ app/src/main/java/com/vsevolodganin/clicktrack/state/MetronomeScreenState.kt
@@ -7,6 +7,6 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class MetronomeScreenState(
     val bpm: BeatsPerMinute,
-    val progress: Float?,
+    val progress: Double?,
     val isPlaying: Boolean,
 ) : Parcelable
